异步编排下订单流程，设置了合理的线程池数量。

+++ api-commons/src/main/java/com/wj/dto/OrderItemDTO.java
@@ -17,7 +17,7 @@ import java.math.BigDecimal;
 public class OrderItemDTO {
     private Long id; //产品id
 
-    private Integer count; //购买数量
+    private Long count; //购买数量
 
     private BigDecimal price; //产品价格
 }
